# Little performance improvements