--- a/src/server/NextApi.Server/Base/NextApiUtils.cs
+++ b/src/server/NextApi.Server/Base/NextApiUtils.cs
@@ -1,4 +1,5 @@
-using System;
+﻿using System;
+using System.Globalization;
 using System.Linq;
 
 namespace NextApi.Server.Base
@@ -14,7 +15,7 @@ namespace NextApi.Server.Base
             {
                 null => throw new ArgumentNullException(nameof(input)),
                 "" => throw new ArgumentException($@"{nameof(input)} cannot be empty", nameof(input)),
-                _ => (input.First().ToString().ToUpper() + input.Substring(1))
+                _ => char.ToUpper(input[0], CultureInfo.InvariantCulture) + input.Substring(1)
             };
         }