Here is benchmark of FirstCharToUpper improvement.

```csharp
        [Params("shortString", "1itStartsWithNumber", ".itStartsWithSymbol")]
        public string input;

        [Benchmark]
        public void OldFirstCharToUpper()
        {
            _ = input switch
            {
                null => throw new ArgumentNullException(nameof(input)),
                "" => throw new ArgumentException($@"{nameof(input)} cannot be empty", nameof(input)),
                _ => (input.First().ToString().ToUpper() + input.Substring(1))
            };
        }

        [Benchmark]
        public void NewFirstCharToUpper()
        {
            _ = input switch
            {
                null => throw new ArgumentNullException(nameof(input)),
                "" => throw new ArgumentException($"{nameof(input)} cannot be empty", nameof(input)),
                _ => char.ToUpper(input[0], CultureInfo.InvariantCulture) + input.Substring(1)
            };
        }

        [Benchmark]
        public void NewFirstCharToUpper_First_Instead_of_Index()
        {
            _ = input switch
            {
                null => throw new ArgumentNullException(nameof(input)),
                "" => throw new ArgumentException($"{nameof(input)} cannot be empty", nameof(input)),
                _ => char.ToUpper(input.First(), CultureInfo.InvariantCulture) + input.Substring(1)
            };
        }
```


``` ini

BenchmarkDotNet=v0.13.0, OS=Windows 10.0.19042.985 (20H2/October2020Update)
Intel Core i5-7400 CPU 3.00GHz (Kaby Lake), 1 CPU, 4 logical and 4 physical cores
.NET SDK=5.0.300
  [Host]     : .NET 5.0.6 (5.0.621.22011), X64 RyuJIT
  DefaultJob : .NET 5.0.6 (5.0.621.22011), X64 RyuJIT


```
|                                     Method |               input |     Mean |    Error |   StdDev |
|------------------------------------------- |-------------------- |---------:|---------:|---------:|
|                        **OldFirstCharToUpper** | **.itStartsWithSymbol** | **78.09 ns** | **1.522 ns** | **1.423 ns** |
|                        NewFirstCharToUpper | .itStartsWithSymbol | 39.39 ns | 0.422 ns | 0.374 ns |
| NewFirstCharToUpper_First_Instead_of_Index | .itStartsWithSymbol | 68.33 ns | 0.525 ns | 0.466 ns |
|                        **OldFirstCharToUpper** | **1itStartsWithNumber** | **77.33 ns** | **0.598 ns** | **0.559 ns** |
|                        NewFirstCharToUpper | 1itStartsWithNumber | 39.38 ns | 0.134 ns | 0.112 ns |
| NewFirstCharToUpper_First_Instead_of_Index | 1itStartsWithNumber | 69.24 ns | 0.487 ns | 0.432 ns |
|                        **OldFirstCharToUpper** |         **shortString** | **89.85 ns** | **1.403 ns** | **1.313 ns** |
|                        NewFirstCharToUpper |         shortString | 38.43 ns | 0.760 ns | 1.138 ns |
| NewFirstCharToUpper_First_Instead_of_Index |         shortString | 68.09 ns | 1.345 ns | 2.248 ns |


And for netcore3.1

``` ini

BenchmarkDotNet=v0.13.0, OS=Windows 10.0.19042.985 (20H2/October2020Update)
Intel Core i5-7400 CPU 3.00GHz (Kaby Lake), 1 CPU, 4 logical and 4 physical cores
.NET SDK=5.0.300
  [Host]     : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT
  DefaultJob : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT


```
|                                     Method |               input |     Mean |    Error |   StdDev |
|------------------------------------------- |-------------------- |---------:|---------:|---------:|
|                        **OldFirstCharToUpper** | **.itStartsWithSymbol** | **81.35 ns** | **0.359 ns** | **0.336 ns** |
|                        NewFirstCharToUpper | .itStartsWithSymbol | 43.65 ns | 0.204 ns | 0.191 ns |
| NewFirstCharToUpper_First_Instead_of_Index | .itStartsWithSymbol | 74.83 ns | 0.414 ns | 0.387 ns |
|                        **OldFirstCharToUpper** | **1itStartsWithNumber** | **84.76 ns** | **1.725 ns** | **2.736 ns** |
|                        NewFirstCharToUpper | 1itStartsWithNumber | 45.48 ns | 0.934 ns | 1.453 ns |
| NewFirstCharToUpper_First_Instead_of_Index | 1itStartsWithNumber | 71.30 ns | 0.639 ns | 0.598 ns |
|                        **OldFirstCharToUpper** |         **shortString** | **95.23 ns** | **1.334 ns** | **1.248 ns** |
|                        NewFirstCharToUpper |         shortString | 41.66 ns | 0.667 ns | 0.557 ns |
| NewFirstCharToUpper_First_Instead_of_Index |         shortString | 69.55 ns | 0.388 ns | 0.344 ns |